% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search_url.R
\name{create_search_url}
\alias{create_search_url}
\title{Creates URLs for a Domain search}
\usage{
create_search_url(
  sale_type = c("rent", "sold-listings"),
  suburbs,
  beds = NULL,
  baths = NULL,
  price = NULL,
  exclude = T,
 
    ua = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36"
)
}
\arguments{
\item{sale_type}{'rent' to search for rentals or 'sold-listings' to search
for recent sales.}

\item{suburbs}{A vector of suburbs with the format
'suburb-name-state-postcode'.}

\item{beds}{A (optional) 2-element vector of minimum and maximum bedrooms.
c(NA, NA) means any number of bedrooms, c(1, 2) means 1-2 bedrooms and c(2,
NA) means 2 or more bedrooms.}

\item{baths}{A (optional) 2-element vector of minimum and maximum bathrooms
with the same format as the number of bedrooms parameter.}

\item{price}{A (optional) 2-element vector of minimum and maximum rents or
prices with the same format as the number of bedrooms parameter. For rental
searches, provide the rent per week, and for sold listings provide the sale
price in dollars.}

\item{exclude}{If TRUE (default), excludes properties with a deposit already
taken from rental searches and excludes properites with price withheld from
sales searches. If FALSE, includes all results.}

\item{ua}{A valid user agent for rvest (optional).}
}
\value{
A vector of URLs.
}
\description{
Creates URLs given parameters that is used when searching properties on
domain.com.au. If the number of results exceeds a page, creates a vector of
URLs that points to every results page.
}
