% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_property_details.R
\name{get_property_details}
\alias{get_property_details}
\title{Scrapes details of properties}
\usage{
get_property_details(urls, get_sale = F)
}
\arguments{
\item{urls}{A vector of property listing URLs.}

\item{get_sale}{If FALSE, treats listings as rental listings. If TRUE, treats
listings as sale listings and gets sale date and method in addition to
existing parameters.}
}
\value{
A table of properties with details on address, asking price,
  bedrooms, bathrooms, parking spaces, property features and description.
}
\description{
Scrapes details of properties given a vector of URLS of their listings and
simplifies the results into a table of property characteristics.
}
